// src/interfaces/InputProps/InputProps.tsx
export interface InputProps {
    type?: string;
    text?: string;
    color?: string;
    fontSize?: string;
    fontWeight?: string;
    fontStyle?: string;
    textTransform?: string;
    fontFamily?: string;
    textDecoration?: string;
    whiteSpace?: string;
    textAlign?: string;
    width?: string;
    height?: string;
    borderRadius?: string;
    bgColor?: string;
    hoverColor?: string;
    padding?: string;
    border?: string;
    hoverBgColor?: string;
    borderFocus?: string;
    focusBgColor?: string;
    search?: boolean;
    monitoring?: boolean;
}